Add test for subprefix hijack with custom announcements and ROAs
This ensures that ROAs are merged with announcements correctly
