+++ bgpy/tests/engine_tests/engine_test_configs/internals/__init__.py
@@ -3,6 +3,7 @@ from .internal_config_001 import internal_config_001
 from .internal_config_002 import internal_config_002
 from .internal_config_003 import internal_config_003
 from .internal_config_004 import internal_config_004
+from .internal_config_005 import internal_config_005
 
 internal_configs = [
     internal_config_000,
@@ -10,6 +11,7 @@ internal_configs = [
     internal_config_002,
     internal_config_003,
     internal_config_004,
+    internal_config_005,
 ]
 
 __all__ = ["internal_configs"]
