+++ bgpy/tests/engine_tests/engine_test_outputs/internal_005/metrics_gt.csv
@@ -1,0 +1,10 @@
+scenario_cls,AdoptingPolicyCls,BasePolicyCls,PolicyCls,outcome_type,as_group,outcome,percent_adopt,propagation_round,value,yerr,scenario_config_label
+CustomScenario,PseudoBGP,BGP,BGP,DATA,all_wout_ixps,ATTACKER_SUCCESS,0,0,0.0,0.0,
+CustomScenario,PseudoBGP,BGP,ROV,DATA,all_wout_ixps,ATTACKER_SUCCESS,0,0,0.0,0.0,
+CustomScenario,PseudoBGP,BGP,Policy,DATA,all_wout_ixps,ATTACKER_SUCCESS,0,0,0.0,0.0,
+CustomScenario,PseudoBGP,BGP,BGP,DATA,all_wout_ixps,VICTIM_SUCCESS,0,0,100.0,0.0,
+CustomScenario,PseudoBGP,BGP,ROV,DATA,all_wout_ixps,VICTIM_SUCCESS,0,0,100.0,0.0,
+CustomScenario,PseudoBGP,BGP,Policy,DATA,all_wout_ixps,VICTIM_SUCCESS,0,0,100.0,0.0,
+CustomScenario,PseudoBGP,BGP,BGP,DATA,all_wout_ixps,DISCONNECTED,0,0,0.0,0.0,
+CustomScenario,PseudoBGP,BGP,ROV,DATA,all_wout_ixps,DISCONNECTED,0,0,0.0,0.0,
+CustomScenario,PseudoBGP,BGP,Policy,DATA,all_wout_ixps,DISCONNECTED,0,0,0.0,0.0,
