+++ bgpy/tests/engine_tests/engine_test_outputs/internal_005/outcomes_gt.yaml
@@ -1,0 +1,5 @@
+1: 1
+2: 1
+3: 1
+666: 0
+777: 1
